Added Dependencies and Constructor
This is a long refactoring and the Class will not compile until
numerous dependcies can be regulated> Therefore although Refactoring
should be move compile test in this case.It seems to be move search for
dependencies.. add any found dependencies. Eclipse shows when the error
is found. But I also don't want to go too far just in case I need to
back track to solve future problems