Removed Method GetSlaughterTension from ServerPlayer
Moved Maethod GetSlaughterTension from ServerPlayer to
ServerPlayerCombat. The methods now in ServerPlayer whose visibilty was
changed from private to protected really belong either in the
ServerPlayerCombat Class in Classes exteneded from ServerPlayerCombat.
This one method move was to see if there were dependencies involved
when moving the class. There were no errors reported so extracting the
rest of the methods should not be too difficult